CPPSourceParser
